refactor telemetry code into a separate file
